Replacing r2d2 syncronous DB pool with async deadpool.
- Removes blocking, makes all connection grabs async.
- Fixes #1684
- Note: clippy::needless_borrow is currently broken.

+++ crates/utils/src/settings/structs.rs
@@ -57,7 +57,7 @@ pub struct DatabaseConfig {
   #[default("lemmy")]
   pub(super) database: String,
   #[default(5)]
-  pub pool_size: u32,
+  pub pool_size: usize,
 }
 
 #[derive(Debug, Deserialize, Clone)]
